requirements-unit-tests.txt
pytest==3.10.1
pytest-aiohttp==0.3.0
asynctest==0.12.2
factory_boy==2.11.0
PyHamcrest==1.9.0